+++ validator/src/commands/manage_block_production/mod.rs
@@ -24,13 +24,12 @@ pub struct ManageBlockProductionArgs {
 
 impl FromClapArgMatches for ManageBlockProductionArgs {
     fn from_clap_arg_match(matches: &ArgMatches) -> Result<Self> {
+        let block_production_method =
+            value_t!(matches, "block_production_method", BlockProductionMethod).unwrap_or_default();
+        block_production_method.warn_if_deprecated_value();
+
         Ok(ManageBlockProductionArgs {
-            block_production_method: value_t!(
-                matches,
-                "block_production_method",
-                BlockProductionMethod
-            )
-            .unwrap_or_default(),
+            block_production_method,
             transaction_structure: value_t!(matches, "transaction_struct", TransactionStructure)
                 .unwrap_or_default(),
             num_workers: value_t!(matches, "block_production_num_workers", NonZeroUsize)