+++ validator/src/commands/run/execute.rs
@@ -951,6 +951,9 @@ pub fn execute(
         bam_url,
         disable_multicast_shred_check: matches.is_present("disable_multicast_shred_check"),
     };
+    validator_config
+        .block_production_method
+        .warn_if_deprecated_value();
 
     let vote_account = pubkey_of(matches, "vote_account").unwrap_or_else(|| {
         if !validator_config.voting_disabled {